feat(frontend): rename folder
